src/minimal_example.py
import grpc
from google.protobuf import wrappers_pb2

def main():
    """
    Minimal example demonstrating how to set up a gRPC client connection.
    Note: This is just a demonstration - no actual server connection will be made.
    """
    # Example connection parameters
    host = "localhost"
    port = 50051
    
    # Create an insecure channel (for demonstration only)
    channel = grpc.insecure_channel(f"{host}:{port}")
    
    try:
        # Example of creating a protobuf message
        string_value = wrappers_pb2.StringValue(value="example_case_id")
        print(f"Created StringValue protobuf: {string_value}")
        
        # Example of how the full client would work
        print("\nIn a complete setup, you would:")
        print("1. Connect to a running gRPC server")
        print("2. Make RPC calls to get Salesforce case details")
        print("3. Process the response data")
        
    except Exception as e:
        print(f"Error occurred: {e}")
    finally:
        channel.close()

if __name__ == "__main__":
    main() 